requirements.txt
opencv-python
dlib
imutils
scipy
face_recognition
numpy
pyodbc
pillow
